refactor(runner): split schema.test.ts into focused test files
Split the 4,271-line schema.test.ts into 6 focused files:
- schema-examples.test.ts: example-based schema tests
- schema-basic.test.ts: basic types, references, schema references, key navigation
- schema-anyof.test.ts: anyOf support including array anyOf
- schema-defaults.test.ts: default value handling
- schema-streams.test.ts: stream support and running promises
- schema-links.test.ts: link resolution, cross-space links, redirect links, toCell behavior
